LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/origin-keyed-agent-clusters/getter-special-cases/sandboxed-same-origin-iframe-no.https-expected.txt


FAIL originAgentCluster must equal false assert_equals: expected (boolean) false but got (undefined) undefined
